+++ backend/connection/connection_subprocess.py
@@ -29,7 +29,9 @@ from custom_types.error import ErrorDict
 from filters import FilterDict
 from filters.filter_data_dict import FilterDataDict
 from custom_types.message import MessageDict
+from session.data.participant import ParticipantDict
 from session.data.participant.participant_summary import ParticipantSummaryDict
+from users.user import User
 
 
 class ConnectionSubprocess(ConnectionInterface):
@@ -68,17 +70,17 @@ class ConnectionSubprocess(ConnectionInterface):
     _local_description: RTCSessionDescription | None
 
     def __init__(
-        self,
-        offer: RTCSessionDescriptionDict,
-        message_handler: Callable[[MessageDict], Coroutine[Any, Any, None]],
-        log_name_suffix: str,
-        config: Config,
-        audio_filters: list[FilterDict],
-        video_filters: list[FilterDict],
-        audio_group_filters: list[FilterDict],
-        video_group_filters: list[FilterDict],
-        filter_api: FilterAPI,
-        record_data: tuple,
+            self,
+            offer: RTCSessionDescriptionDict,
+            message_handler: Callable[[MessageDict], Coroutine[Any, Any, None]],
+            log_name_suffix: str,
+            config: Config,
+            audio_filters: list[FilterDict],
+            video_filters: list[FilterDict],
+            audio_group_filters: list[FilterDict],
+            video_group_filters: list[FilterDict],
+            filter_api: FilterAPI,
+            record_data: tuple,
     ):
         """Create new ConnectionSubprocess.
 
@@ -143,15 +145,15 @@ class ConnectionSubprocess(ConnectionInterface):
         return self._state
 
     async def create_subscriber_proposal(
-        self, participant_summary: ParticipantSummaryDict | str | None,
-        subscriber=None
+            self, participant_summary: ParticipantSummaryDict | str | None,
+            subscriber: ParticipantDict = None
     ) -> ConnectionOfferDict:
         # For docstring see ConnectionInterface or hover over function declaration
         # Send command and wait for response.
         data = dict()
         data['participant_summary'] = participant_summary
         if subscriber is not None:
-            data['subscriber'] = subscriber.get_participant_data()
+            data['subscriber'] = subscriber
         offer = await self._send_command_wait_for_response("CREATE_PROPOSAL", data)
         return offer
 
@@ -160,7 +162,7 @@ class ConnectionSubprocess(ConnectionInterface):
         await self._send_command("ADD_ICE_CANDIDATE", candidate)
 
     async def handle_subscriber_offer(
-        self, offer: ConnectionOfferDict
+            self, offer: ConnectionOfferDict
     ) -> ConnectionAnswerDict:
         # For docstring see ConnectionInterface or hover over function declaration
         # Send command and wait for response.
@@ -168,7 +170,7 @@ class ConnectionSubprocess(ConnectionInterface):
         return answer
 
     async def handle_subscriber_add_ice_candidate(
-        self, candidate: AddIceCandidateDict
+            self, candidate: AddIceCandidateDict
     ):
         # For docstring see ConnectionInterface or hover over function declaration
         await self._send_command("ADD_SUBSCRIBER_ICE_CANDIDATE", candidate)
@@ -216,13 +218,13 @@ class ConnectionSubprocess(ConnectionInterface):
         await self._send_command("SET_AUDIO_FILTERS", filters)
 
     async def set_video_group_filters(
-        self, group_filters: list[FilterDict], ports: list[tuple[int, int]]
+            self, group_filters: list[FilterDict], ports: list[tuple[int, int]]
     ) -> None:
         # For docstring see ConnectionInterface or hover over function declaration
         await self._send_command("SET_VIDEO_GROUP_FILTERS", (group_filters, ports))
 
     async def set_audio_group_filters(
-        self, group_filters: list[FilterDict], ports: list[tuple[int, int]]
+            self, group_filters: list[FilterDict], ports: list[tuple[int, int]]
     ) -> None:
         # For docstring see ConnectionInterface or hover over function declaration
         await self._send_command("SET_AUDIO_GROUP_FILTERS", (group_filters, ports))
@@ -451,19 +453,19 @@ class ConnectionSubprocess(ConnectionInterface):
         await queue.put(data)
 
     async def _send_command_wait_for_response(
-        self,
-        command: str,
-        data: str
-        | int
-        | float
-        | dict
-        | None
-        | tuple
-        | ParticipantSummaryDict
-        | ConnectionOfferDict
-        | MessageDict,
-        timeout: int | None = None,
-        retries: int = 1,
+            self,
+            command: str,
+            data: str
+                  | int
+                  | float
+                  | dict
+                  | None
+                  | tuple
+                  | ParticipantSummaryDict
+                  | ConnectionOfferDict
+                  | MessageDict,
+            timeout: int | None = None,
+            retries: int = 1,
     ) -> dict | None:
         """Send a command including unique command_nr and wait for response.
 
@@ -518,19 +520,19 @@ class ConnectionSubprocess(ConnectionInterface):
             return answer
 
     async def _send_command(
-        self,
-        command: str,
-        data: str
-        | int
-        | float
-        | dict
-        | None
-        | tuple
-        | ParticipantSummaryDict
-        | ConnectionOfferDict
-        | MessageDict
-        | list,
-        command_nr: int = -1,
+            self,
+            command: str,
+            data: str
+                  | int
+                  | float
+                  | dict
+                  | None
+                  | tuple
+                  | ParticipantSummaryDict
+                  | ConnectionOfferDict
+                  | MessageDict
+                  | list,
+            command_nr: int = -1,
     ) -> None:
         """Send command to subprocess via stdin.
 
@@ -567,16 +569,16 @@ ConnectionInterface.register(ConnectionSubprocess)
 
 
 async def connection_subprocess_factory(
-    offer: RTCSessionDescription,
-    message_handler: Callable[[MessageDict], Coroutine[Any, Any, None]],
-    log_name_suffix: str,
-    config: Config,
-    audio_filters: list[FilterDict],
-    video_filters: list[FilterDict],
-    audio_group_filters: list[FilterDict],
-    video_group_filters: list[FilterDict],
-    filter_api: FilterAPI,
-    record_data: tuple,
+        offer: RTCSessionDescription,
+        message_handler: Callable[[MessageDict], Coroutine[Any, Any, None]],
+        log_name_suffix: str,
+        config: Config,
+        audio_filters: list[FilterDict],
+        video_filters: list[FilterDict],
+        audio_group_filters: list[FilterDict],
+        video_group_filters: list[FilterDict],
+        filter_api: FilterAPI,
+        record_data: tuple,
 ) -> Tuple[RTCSessionDescription, ConnectionSubprocess]:
     """Instantiate new ConnectionSubprocess.
